# openapi/openapi3-v1.yaml
...
openapi: 3.0.3
info:
  title: Contract Manager
  description: Contract Manager HTTP APIs for interaction with the Dataspace connector.
  version: '1.1'
  license:
    name: European Union Public License (EUPL) 1.2
    url: https://interoperable-europe.ec.europa.eu/sites/default/files/custom-page/attachment/eupl_v1.2_en.pdf
  contact:
    name: Simpl Programme
    email: cnect-simpl@ec.europa.eu
    url: https://simpl-programme.ec.europa.eu/
servers:
  - url: https://localhost:8085/contract/v1
    description: Contract Manager
paths:
...
  "/credentials/agreements/{contractAgreementId}/definitions/{contractDefinitionId}/issue":
    post:
      tags:
        - verifiable-credentials
      summary: Issue Verifiable credential
      description: Issue Verifiable credential. Start the process of signing contract.
      operationId: contract#CreateVerifiableCredential
      requestBody:
        description: Request body
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ContractAgreementCreate"
            example:
              contractNegotiationId: 0ec61c92-677b-4e97-aef1-ed542e37183b
              assetId: MyResource1
              providerId: FirstProvider
              consumerId: Cons2000
              contractOfferId: 32c61c92-677b-4e97-aef1-ed542e371234
        required: false
      parameters:
        - name: contractAgreementId
          in: path
          description: Contract Agreement Id
          required: true
          schema:
            type: string
            format: uuid
          example: 0ec61c92-677b-4e97-aef1-ed542e37183b
        - name: contractDefinitionId
          in: path
          description: Contract Definition Id
          required: true
          schema:
            type: string
          example: 0ec61c92-677b-4e97-aef1-ed542e37183b
      responses:
        '201':
          description: OK response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ConfirmationResponse"
              example:
                contractAgreementId: 677186f6-1179-4cf1-b845-68d1c00c185f
                contractDefinitionId: 777186f6-1179-4cf1-b845-68d1c00c185f
                status: INITIATED
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 400
                title: Bad Request
                detail: Invalid body content
                instance: "/contract/v1/credentials/agreements/7e822f42-f17c-4849-a56f-fbdb5209cd17/definitions/def1/issue"
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 401
                title: Unauthorized
                detail: Invalid API Key
        '405':
          description: Method Not allowed
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 405
                title: Method Not Allowed
                instance: "/contract/v1/credentials/agreements/7e822f42-f17c-4849-a56f-fbdb5209cd17/definitions/def1/issue"
        '500':
          description: Server Error
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 500
                title: Server Error
                detail: Server Error
  "/credentials/agreements/{contractAgreementId}/definitions/{contractDefinitionId}/create":
    post:
      tags:
        - verifiable-credentials
      summary: Create contract agreement data
      description: Initiate process of collecting all the data for contract presentation
      operationId: contract#CreateContractData
      requestBody:
        description: Request body
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ContractAgreementCreate"
            example:
              contractNegotiationId: 0ec61c92-677b-4e97-aef1-ed542e37183b
              assetId: MyResource1
              providerId: FirstProvider
              consumerId: Cons2000
              contractOfferId: 32c61c92-677b-4e97-aef1-ed542e371234
        required: true
      parameters:
        - name: contractAgreementId
          in: path
          description: Contract Agreement Id
          required: true
          schema:
            type: string
            format: uuid
          example: 0ec61c92-677b-4e97-aef1-ed542e37183b
        - name: contractDefinitionId
          in: path
          description: Contract Definition Id
          required: true
          schema:
            type: string
          example: 0ec61c92-677b-4e97-aef1-ed542e37183b
      responses:
        '201':
          description: OK response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ConfirmationResponse"
              example:
                contractAgreementId: 677186f6-1179-4cf1-b845-68d1c00c185f
                contractDefinitionId: 777186f6-1179-4cf1-b845-68d1c00c185f
                status: INITIATED
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 400
                title: Bad Request
                detail: Invalid body content
                instance: "/contract/v1/credentials/agreements/7e822f42-f17c-4849-a56f-fbdb5209cd17/definitions/def1/create"
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 401
                title: Unauthorized
                detail: Invalid API Key
        '405':
          description: Method Not allowed
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 405
                title: Method Not Allowed
                instance: "/contract/v1/credentials/agreements/7e822f42-f17c-4849-a56f-fbdb5209cd17/definitions/def1/create"
        '415':
          description: Unsupported Media Type
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 415
                title: Unsupported Media Type
                instance: "/contract/v1/credentials/agreements/7e822f42-f17c-4849-a56f-fbdb5209cd17/definitions/def1/create"
        '500':
          description: Server Error
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 500
                title: Server Error
                detail: Server Error
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    ConfirmationResponse:
      type: object
      required:
        - contractAgreementId
        - contractDefinitionId
        - status
      properties:
        contractAgreementId:
          type: string
          format: uuid
          description: ID of contract agreement
        contractDefinitionId:
          type: string
          description: ID of contract definition
        status:
          type: string
          description: Status of the process, in initiated state.
          example: TERMINATED
          format: string
          enum:
            - INITIATED
            - CREATED
            - FINALIZING
            - FINALIZED
            - TERMINATED
      example:
        contractAgreementId: 677186f6-1179-4cf1-b845-68d1c00c185f
        contractDefinitionId: 777186f6-1179-4cf1-b845-68d1c00c185f
        status: FINALIZING
...
    ContractAgreementCreate:
      type: object
      properties:
        contractNegotiationId:
          type: string
          example: "C1-12"
          format: string
        assetId:
          type: string
          example: "C1-12"
          format: string
        providerId:
          type: string
          example: "provider1"
          format: string
        consumerId:
          type: string
          example: "consumer1000"
          format: string
        contractOfferId:
          type: string
          example: "offerId123"
          format: string
      example:
        contractNegotiationId: C1-12
        assetId: As5
        providerId: PR1
        consumerId: CNS44
        contractOfferId: C1-12-33
...
    Problem:
      description: A Problem Details object (RFC 9457). Additional properties specific
        to the problem type may be present.
      type: object
      properties:
        type:
          type: string
          format: uri
          description: An absolute URI that identifies the problem type
          default: about:blank
        title:
          type: string
          description: A short summary of the problem type. Written in English and
            readable for engineers (usually not suited for non technical stakeholders
            and not localized).
          example: Service Unavailable
        status:
          type: integer
          format: int32
          description: The HTTP status code generated by the origin server for this
            occurrence of the problem.
          minimum: 400
          maximum: 600
          exclusiveMaximum: true
          example: 503
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of
            the problem
        instance:
          type: string
          format: uri
          description: An absolute URI that identifies the specific occurrence of
            the problem. It may or may not yield further information if dereferenced.
...
security:
  - ApiKeyAuth: []